correct the format

+++ src/bus/iic/hardware_iic_2.cpp
@@ -2,7 +2,7 @@
  * @Description: None
  * @Author: LILYGO_L
  * @Date: 2025-02-13 15:04:49
- * @LastEditTime: 2025-08-09 11:58:17
+ * @LastEditTime: 2025-08-20 14:45:32
  * @License: GPL 3.0
  */
 #include "hardware_iic_2.h"
@@ -11,7 +11,7 @@ namespace Cpp_Bus_Driver
 {
     bool Hardware_Iic_2::begin(uint32_t freq_hz, uint16_t address)
     {
-        if (freq_hz == DEFAULT_CPP_BUS_DRIVER_VALUE)
+        if (freq_hz == static_cast<uint32_t>(DEFAULT_CPP_BUS_DRIVER_VALUE))
         {
             freq_hz = DEFAULT_CPP_BUS_DRIVER_IIC_FREQ_HZ;
         }
